comm -> createbean update
