+++ JavaJourney/src/main/java/web/dao/mapper/CreateBeanDaoMapper.xml
@@ -1,0 +1,7 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<!DOCTYPE mapper
+    PUBLIC "-//mybatis.org//DTD Mapper 3.0//EN"
+    "https://mybatis.org/dtd/mybatis-3-mapper.dtd">
+<mapper namespace="web.dao.face.CreateBeanDao">
+
+</mapper>
