+++ JavaJourney/src/main/java/web/service/face/CreateBeanService.java
@@ -1,0 +1,14 @@
+package web.service.face;
+
+import web.dto.Bean;
+
+public interface CreateBeanService {
+	
+	/**
+	 * 원두(구독) 글 삽입 로직
+	 * DTO 중 파일(이미지)자료형 MultipartFile beanInfo;
+	 * @param bean - beanName, origin, beanComm, beanPrice, businessNo
+	 */
+	public void insertBean(Bean bean);
+
+}
