+++ JavaJourney/src/main/java/web/service/impl/CommunityServiceImpl.java
@@ -5,7 +5,6 @@ import org.springframework.stereotype.Service;
 
 import lombok.extern.slf4j.Slf4j;
 import web.dao.face.CommunityDao;
-import web.dto.Bean;
 import web.service.face.CommunityService;
 
 @Service
@@ -13,11 +12,6 @@ import web.service.face.CommunityService;
 public class CommunityServiceImpl implements CommunityService {
 	
 	@Autowired private CommunityDao dao;
-//	@Autowired private ServletContext context;
 	
-	@Override
-	public void insertBean(Bean bean) {
-		log.info("bean: {}",bean);
-		dao.insertBeanData(bean);
-	}
+
 }
